requirements.txt
grpcio
grpcio-tools
-e .

pytest
pytest-cov
flake8
autoflake
pep8
autopep8
requests
backports.tempfile
